Fixing startup test

+++ window/src/main/java/com/haines/ml/rce/window/WindowConfig.java
@@ -2,7 +2,15 @@ package com.haines.ml.rce.window;
 
 public interface WindowConfig {
 
-	int getWindowPeriod();
-
+	/**
+	 * Returns the milliseconds that each window represents
+	 * @return
+	 */
+	long getWindowPeriod();
+	
+	/**
+	 * Number of windows to hold events for
+	 * @return
+	 */
 	int getNumWindows();
 }
